// src/types/buttons.ts
export enum SortArrayButtons {
  BubbleSort = "bubbleSort",
  SelectionSort = "selectionSort",
  Ascending = "ascending",
  Descending = "descending",
  NewArray = "New Array",
}

export enum StackAndQueueButtons {
  Add = "add",
  Remove = "remove",
  Clear = "clear",
}

export enum LinkedListButtons {
  AddToHead = "addToHead",
  AddToEnd = "addToEnd",
  RemoveFromEnd = "removeFromEnd",
  RemoveFromHead = "removeFromHead",
  InsertByIndex = "insertByIndex",
  ExtractByIndex = "extractByIndex",
}